backend/requirements.txt
fastapi
uvicorn
openai
langchain
pypdf
chromadb
tiktoken
python-dotenv